finish kangaskhan and clean event flag
