+++ ld_script.txt
@@ -97,8 +97,6 @@ SECTIONS {
         src/code_8012A18_1.o(.text);
         src/felicity_bank.o(.text);
         src/kangaskhan_storage.o(.text);
-        asm/kangaskhan_storage.o(.text);
-        src/kangaskhan_storage_1.o(.text);
         src/kecleon_items.o(.text);
         asm/kecleon_items.o(.text);
         src/kecleon_items_1.o(.text);
